将记录区块对应seq 和 推送注册相关的代码分开
